Working

+++ pyclipper/ytdl.py
@@ -1,4 +1,3 @@
-import logging
 import os
 from urllib.parse import quote
 import ffmpeg
@@ -7,11 +6,11 @@ from durations import Duration
 from moviepy.video.io.ffmpeg_tools import ffmpeg_extract_subclip
 from requests.compat import urljoin
 
-logger = logging.getLogger(__name__)
+from pyclipper.config import Config
+from pyclipper.timestamp import VideoTimestamp
 
-ngrok = "https://4578c91c.ngrok.io"
-# TODO: ENV config
-base = ngrok
+c = Config()
+
 
 class EndTimeNeededError(Exception):
     def __init__(self):
@@ -23,30 +22,29 @@ def download_and_trim(video_identifier, start, end=None):
     if end is None:
         raise EndTimeNeededError()
 
-    start_seconds = Duration(start).to_seconds()
-    end_seconds = Duration(end).to_seconds()
+    if not isinstance(video_identifier, str):
+        return
 
     videos = "pyclipper/assets"
-    full = f"{videos}/full"
-    clips = f"{videos}/clips"
+    full_video_path = f"{videos}/full"
+    clips_path = f"{videos}/clips"
     extension = ".mp4"
 
     template = "pyclipper/assets/full/%(title)s-%(id)s.%(ext)s"
-    print(template)
 
     ydl_opts = {"outtmpl": template,
-                # "format": "137"
+                "format": "mp4"
                 }
 
     with youtube_dl.YoutubeDL(ydl_opts) as ydl:
-        info = ydl.extract_info(video_identifier, False)
+        info = ydl.extract_info(video_identifier, download=False)
         title = info['title']
         video_id = info['id']
         ext = info['ext']
-        video = f"{full}/{title}-{video_id}.{ext}"
-        clip_name = f"{title}-{video_id}-s{start_seconds}-e{end_seconds}{extension}"
-        target = os.path.join(clips, "buggy" + clip_name)
-        clip_path = os.path.join(clips, clip_name)
+        video = f"{full_video_path}/{title}-{video_id}.{ext}"
+        clip_name = f"{title}-{video_id}-s{start}-e{end}{extension}"
+        target = os.path.join(clips_path, "buggy" + clip_name)
+        clip_path = os.path.join(clips_path, clip_name)
 
         if not os.path.exists(video):
             ydl.download([video_identifier])
@@ -58,10 +56,12 @@ def download_and_trim(video_identifier, start, end=None):
     # freezes and runs until 5 seconds. The second call to ffmpeg.input... fixes this
     # Solution would be to use ffmpeg.input to extract the clip in the first place,
     # but I've got to read more about the library
-    ffmpeg_extract_subclip(video, start_seconds, end_seconds, targetname=target)
-    ffmpeg.input(target).output(clip_path).run()
-    os.remove(target)
-    return urljoin(base, f"clips/{quote(clip_name)}")
+    if not os.path.exists(clip_path):
+        ffmpeg_extract_subclip(video, start, end, targetname=target)
+        ffmpeg.input(target).output(clip_path).run()
+        os.remove(target)
+
+    return urljoin(c.base_url, f"clips/{quote(clip_name)}")
